server/requirements/dist.txt
Django==3.2.14
django-rest-auth==0.9.5
dj-rest-auth==2.2.4
django-allauth==0.44.0
